clean

+++ recur_facto/recur_fact.c
@@ -1,25 +1,20 @@
 #include <stdio.h>
 #include <math.h>
 
-
-long factorial(long num);
-
+unsigned long long  factorial(unsigned long long  num);
 
 
 int main(void)
 {
 
-   for( int i = 0; i <= 20; i++)
-		printf(" factorial %2d!= %ld \n", i, factorial(i));
-		//printf("test");
+   for( unsigned long long  i = 0; i <= 20; i++)
+		printf(" factorial %2d!= %llu \n", i, factorial(i));
 		
 	return 0;
 }
 
 
-
-/* Функция вычисления факториала */
-long factorial(long num)
+unsigned long long factorial(unsigned long long  num)
 {
 	
 	if ( num <= 1 )
